chore(codegen): #1098 split expr.rs — extract write_barrier, helpers, url_helpers, object_literal, array_literal, index submodules (behavior-preserving)
