+++ crates/perry-codegen/src/expr/mod.rs
@@ -36,8 +36,14 @@ use crate::types::{DOUBLE, I1, I32, I64, I8, PTR};
 // `lower_expr` and the foundational types (`FnCtx`, `FlatConstInfo`)
 // remain here. `pub(crate) use` keeps the public surface stable so
 // existing `crate::expr::X` paths resolve unchanged.
+mod array_literal;
 mod channel;
+mod helpers;
 mod i32_fast_path;
+mod index;
+mod object_literal;
+mod url_helpers;
+mod write_barrier;
 
 #[allow(unused_imports)] // ChannelReduction kept reachable for surface stability
 pub(crate) use channel::{
@@ -48,6 +54,18 @@ pub(crate) use i32_fast_path::{
     can_lower_expr_as_i32, is_known_finite, lower_expr_as_i32, try_flat_const_2d_int,
     try_lower_flat_const_index_get,
 };
+pub(crate) use write_barrier::{
+    emit_write_barrier, emit_write_barrier_slot_on_block, lower_stream_super_init,
+};
+pub(crate) use array_literal::lower_array_literal;
+pub(crate) use helpers::{
+    buffer_alias_metadata_suffix, is_global_this_builtin_function_name,
+    is_global_this_builtin_name, lower_js_args_array, proxy_build_args_array, unbox_str_handle,
+    unbox_to_i64,
+};
+pub(crate) use index::lower_index_set_fast;
+pub(crate) use object_literal::lower_object_literal;
+pub(crate) use url_helpers::lower_url_string_getter;
 
 /// Inline NaN-box of a raw heap pointer with `POINTER_TAG`.
 pub(crate) fn nanbox_pointer_inline(blk: &mut LlBlock, ptr_i64: &str) -> String {
@@ -1077,166 +1095,6 @@ impl<'a> FnCtx<'a> {
 
 /// Lower an expression to a raw LLVM `double` value. Returns the string form
 /// of the value (either a `%rN` register or a literal like `42.0`).
-/// Gen-GC Phase C2 helper: emit a write barrier after heap-store sites
-/// when `PERRY_WRITE_BARRIERS=1`. Sites with a precise field/element
-/// address use `js_write_barrier_slot`; opaque helper stores keep using
-/// the compatibility wrapper, which conservatively marks the parent span.
-/// The env gate is read once and OnceLock-cached at codegen time.
-fn emit_write_barrier(ctx: &mut FnCtx<'_>, parent_bits: &str, child_bits: &str) {
-    if !crate::codegen::write_barriers_enabled() {
-        return;
-    }
-    ctx.block()
-        .call_void("js_write_barrier", &[(I64, parent_bits), (I64, child_bits)]);
-}
-
-fn emit_write_barrier_slot_on_block(
-    blk: &mut LlBlock,
-    parent_bits: &str,
-    slot_addr: &str,
-    child_bits: &str,
-) {
-    if !crate::codegen::write_barriers_enabled() {
-        return;
-    }
-    blk.call_void(
-        "js_write_barrier_slot",
-        &[(I64, parent_bits), (I64, slot_addr), (I64, child_bits)],
-    );
-}
-
-/// Issue #562 — `super({ ... })` for `class X extends ReadableStream`,
-/// `WritableStream`, or `TransformStream`. Extracts the underlying
-/// source/sink/transformer callbacks from the inline object literal,
-/// lowers each one (TAG_UNDEFINED for missing fields), and calls the
-/// runtime `*_subclass_init` shim — which allocates the stream registry
-/// handle and stashes it on `this` under `__perry_stream_handle__`.
-///
-/// `kind` is one of `"readable"` / `"writable"` / `"transform"` —
-/// matches the SuperCall arm's `parent_name` switch in expr.rs.
-fn lower_stream_super_init(ctx: &mut FnCtx<'_>, kind: &str, super_args: &[Expr]) -> Result<String> {
-    let undef_lit = double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED));
-
-    // Pre-extract field exprs so we don't hold a borrow across `lower_expr`.
-    let opts_props: Option<Vec<(String, Expr)>> = super_args
-        .first()
-        .and_then(|first| crate::lower_call::extract_options_fields(ctx, first));
-    let qstrat_props: Option<Vec<(String, Expr)>> = super_args
-        .get(1)
-        .and_then(|second| crate::lower_call::extract_options_fields(ctx, second));
-
-    // Lower the canonical callback set per stream kind. Fields not
-    // present (or callable arg shape that isn't an inline literal) fall
-    // back to TAG_UNDEFINED — matches the existing `new ReadableStream
-    // / WritableStream / TransformStream` lowerings in
-    // `lower_call/builtin.rs`.
-    let mut start = undef_lit.clone();
-    let mut pull = undef_lit.clone();
-    let mut cancel = undef_lit.clone();
-    let mut write = undef_lit.clone();
-    let mut close = undef_lit.clone();
-    let mut abort = undef_lit.clone();
-    let mut transform = undef_lit.clone();
-    let mut flush = undef_lit.clone();
-
-    if let Some(props) = opts_props {
-        for (k, vexpr) in &props {
-            match (kind, k.as_str()) {
-                ("readable", "start") => start = lower_expr(ctx, vexpr)?,
-                ("readable", "pull") => pull = lower_expr(ctx, vexpr)?,
-                ("readable", "cancel") => cancel = lower_expr(ctx, vexpr)?,
-                ("writable", "write") => write = lower_expr(ctx, vexpr)?,
-                ("writable", "close") => close = lower_expr(ctx, vexpr)?,
-                ("writable", "abort") => abort = lower_expr(ctx, vexpr)?,
-                ("transform", "transform") => transform = lower_expr(ctx, vexpr)?,
-                ("transform", "flush") => flush = lower_expr(ctx, vexpr)?,
-                _ => {
-                    // Lower for side effects (closure-capture collection,
-                    // string-pool registration, etc.) but discard the value.
-                    let _ = lower_expr(ctx, vexpr)?;
-                }
-            }
-        }
-    } else if let Some(first) = super_args.first() {
-        // Caller passed something that isn't a recognized shape — lower
-        // for side effects so closure analysis stays consistent.
-        let _ = lower_expr(ctx, first)?;
-    }
-
-    let mut hwm = double_literal(1.0);
-    if let Some(qprops) = qstrat_props {
-        for (k, vexpr) in &qprops {
-            if k == "highWaterMark" {
-                hwm = lower_expr(ctx, vexpr)?;
-            } else {
-                let _ = lower_expr(ctx, vexpr)?;
-            }
-        }
-    } else if let Some(second) = super_args.get(1) {
-        let _ = lower_expr(ctx, second)?;
-    }
-
-    // `this` (NaN-boxed pointer) — the runtime shim stashes the handle
-    // on it via `js_object_set_field_by_name`.
-    let this_slot = ctx.this_stack.last().cloned();
-    let this_box = match this_slot {
-        Some(slot) => ctx.block().load(DOUBLE, &slot),
-        None => undef_lit.clone(),
-    };
-
-    let runtime_fn = match kind {
-        "readable" => "js_readable_stream_subclass_init",
-        "writable" => "js_writable_stream_subclass_init",
-        "transform" => "js_transform_stream_subclass_init",
-        _ => unreachable!("lower_stream_super_init: unexpected kind {}", kind),
-    };
-
-    let blk = ctx.block();
-    match kind {
-        "readable" => {
-            blk.call(
-                DOUBLE,
-                runtime_fn,
-                &[
-                    (DOUBLE, &this_box),
-                    (DOUBLE, &start),
-                    (DOUBLE, &pull),
-                    (DOUBLE, &cancel),
-                    (DOUBLE, &hwm),
-                ],
-            );
-        }
-        "writable" => {
-            blk.call(
-                DOUBLE,
-                runtime_fn,
-                &[
-                    (DOUBLE, &this_box),
-                    (DOUBLE, &write),
-                    (DOUBLE, &close),
-                    (DOUBLE, &abort),
-                    (DOUBLE, &hwm),
-                ],
-            );
-        }
-        "transform" => {
-            blk.call(
-                DOUBLE,
-                runtime_fn,
-                &[
-                    (DOUBLE, &this_box),
-                    (DOUBLE, &transform),
-                    (DOUBLE, &flush),
-                    (DOUBLE, &hwm),
-                ],
-            );
-        }
-        _ => unreachable!(),
-    }
-
-    Ok(double_literal(f64::from_bits(crate::nanbox::TAG_UNDEFINED)))
-}
-
 pub(crate) fn lower_expr(ctx: &mut FnCtx<'_>, expr: &Expr) -> Result<String> {
     match expr {
         // -------- Literals --------
@@ -13755,733 +13613,4 @@ pub(crate) fn lower_expr(ctx: &mut FnCtx<'_>, expr: &Expr) -> Result<String> {
             variant_name(other)
         ),
     }
-}
-
-/// Build a NaN-boxed Array JSValue from a slice of Expr arguments.
-fn proxy_build_args_array(ctx: &mut FnCtx<'_>, args: &[Expr]) -> Result<String> {
-    let cap = (args.len() as u32).to_string();
-    let arr = ctx.block().call(I64, "js_array_alloc", &[(I32, &cap)]);
-    let mut current = arr;
-    for a in args {
-        let v = lower_expr(ctx, a)?;
-        current = ctx
-            .block()
-            .call(I64, "js_array_push_f64", &[(I64, &current), (DOUBLE, &v)]);
-    }
-    Ok(current)
-}
-
-/// Build the `, !alias.scope !N, !noalias !M` suffix attached to Buffer
-/// load/store instructions on the GEP fast path. `scope_idx` is the per-
-/// buffer identifier allocated by `Stmt::Let` when a `BufferAlloc` init
-/// is detected. The metadata IDs map to nodes emitted at module level
-/// by `emit_buffer_alias_metadata` (`codegen.rs`):
-///
-/// - `!(201 + idx)` is the alias-scope list containing this buffer's scope
-/// - `!(301 + idx)` is the noalias set listing every *other* buffer's scope
-///
-/// LLVM's LoopVectorizer uses these to prove that loads from one buffer
-/// don't alias stores to another buffer — the fix for the "unsafe
-/// dependent memory operations" vectorization remark on the image_conv
-/// blur kernel (src reads vs dst writes).
-pub(crate) fn buffer_alias_metadata_suffix(scope_idx: u32) -> String {
-    let scope_list = 201 + scope_idx;
-    let noalias_list = 301 + scope_idx;
-    format!(", !alias.scope !{}, !noalias !{}", scope_list, noalias_list)
-}
-
-/// Marshal a vector of already-lowered NaN-boxed args into a stack
-/// alloca'd `[N x double]` and return `(args_ptr, args_len_str)` ready
-/// for an FFI call expecting `(*const f64, usize)`. Empty input returns
-/// `("null", "0")` so the FFI sees a null pointer + 0 length.
-///
-/// Issue #167: the alloca is hoisted to the function entry block via
-/// `alloca_entry_array` so every per-iteration call inside a loop
-/// reuses one stack slot instead of permanently shrinking the stack.
-fn lower_js_args_array(ctx: &mut FnCtx<'_>, lowered_args: &[String]) -> (String, String) {
-    if lowered_args.is_empty() {
-        return ("null".to_string(), "0".to_string());
-    }
-    let n = lowered_args.len();
-    let buf = ctx.func.alloca_entry_array(DOUBLE, n);
-    for (i, v) in lowered_args.iter().enumerate() {
-        let slot = ctx.block().gep(DOUBLE, &buf, &[(I64, &format!("{}", i))]);
-        ctx.block().store(DOUBLE, v, &slot);
-    }
-    let ptr_reg = ctx.block().next_reg();
-    ctx.block().emit_raw(format!(
-        "{} = getelementptr [{} x double], ptr {}, i64 0, i64 0",
-        ptr_reg, n, buf
-    ));
-    (ptr_reg, n.to_string())
-}
-
-/// Unbox a NaN-boxed double into a raw i64 pointer via inline
-/// `bitcast double → i64; and POINTER_MASK_I64`.
-///
-/// **⚠ Use [`unbox_str_handle`] instead when the value may be a JS string.**
-/// The bitcast+mask returns the lower 48 bits, which is the correct
-/// `*ObjectHeader` / `*ArrayHeader` / `*ClosureHeader` for heap pointers
-/// (POINTER_TAG = 0x7FFD, ARRAY_TAG = 0x7FFB, etc.) and the correct
-/// `*StringHeader` for **heap** strings (STRING_TAG = 0x7FFF), but is
-/// **garbage** for short-string-optimization values (SHORT_STRING_TAG =
-/// 0x7FF9), whose lower 48 bits encode the inline length + bytes. Any
-/// runtime function that dereferences the resulting i64 as a
-/// `*StringHeader` (reading `byte_len`, copying the UTF-8 bytes, …) will
-/// segfault or return garbage on SSO inputs.
-///
-/// SSO-vulnerable callsites must route through [`unbox_str_handle`].
-/// Issue #214 lineage: `Array.indexOf`, every `String.prototype.*` method,
-/// `arr.join(sep)`, `obj[dynamicKey]`, `string.match(re)`, crypto digest
-/// inputs, `process.env[name]` — all previously segfaulted on SSO operands
-/// before being routed through the safe helper.
-pub(crate) fn unbox_to_i64(blk: &mut LlBlock, boxed: &str) -> String {
-    let bits = blk.bitcast_double_to_i64(boxed);
-    blk.and(I64, &bits, POINTER_MASK_I64)
-}
-
-/// Built-in constructor / namespace names that the runtime pre-populates
-/// on the globalThis singleton (`populate_global_this_builtins` in
-/// crates/perry-runtime/src/object.rs). Used by codegen to decide whether
-/// `globalThis.<Name>` should route through `js_get_global_this`
-/// (returning the populated backing-object) or fall through to the `0.0`
-/// no-value placeholder. Keep this list in sync with
-/// `GLOBAL_THIS_BUILTIN_CONSTRUCTORS` + `GLOBAL_THIS_BUILTIN_NAMESPACES`
-/// in object.rs — the codegen check and the runtime population together
-/// implement the lodash `runInContext` blocker fix.
-pub(crate) fn is_global_this_builtin_name(name: &str) -> bool {
-    matches!(
-        name,
-        // Constructors (typeof === "function" in spec).
-        "Array"
-            | "Object"
-            | "String"
-            | "Number"
-            | "Boolean"
-            | "Function"
-            | "RegExp"
-            | "Date"
-            | "Error"
-            | "TypeError"
-            | "RangeError"
-            | "SyntaxError"
-            | "ReferenceError"
-            | "EvalError"
-            | "URIError"
-            | "Symbol"
-            | "Promise"
-            | "Map"
-            | "Set"
-            | "WeakMap"
-            | "WeakSet"
-            | "WeakRef"
-            | "Proxy"
-            | "BigInt"
-            | "Uint8Array"
-            | "Int8Array"
-            | "Uint16Array"
-            | "Int16Array"
-            | "Uint32Array"
-            | "Int32Array"
-            | "Float32Array"
-            | "Float64Array"
-            | "Uint8ClampedArray"
-            | "BigInt64Array"
-            | "BigUint64Array"
-            | "ArrayBuffer"
-            | "SharedArrayBuffer"
-            | "DataView"
-            | "TextEncoder"
-            | "TextDecoder"
-            | "URL"
-            | "URLSearchParams"
-            | "AbortController"
-            | "AbortSignal"
-            | "FormData"
-            | "Headers"
-            | "Request"
-            | "Response"
-            | "FinalizationRegistry"
-            // Namespaces (typeof === "object" in spec).
-            | "Math"
-            | "JSON"
-            | "Reflect"
-    )
-}
-
-/// Subset of `is_global_this_builtin_name` whose `typeof` is `"function"`
-/// in spec (constructors). Used by the `Expr::TypeOf` short-circuit so
-/// `typeof globalThis.Array === "function"`. Math/JSON/Reflect are
-/// namespaces — they keep `typeof === "object"` via the existing match
-/// arms.
-pub(crate) fn is_global_this_builtin_function_name(name: &str) -> bool {
-    is_global_this_builtin_name(name) && !matches!(name, "Math" | "JSON" | "Reflect")
-}
-
-/// SSO-safe variant of `unbox_to_i64` for NaN-boxed string operands.
-///
-/// The plain `unbox_to_i64(bitcast double → i64; and POINTER_MASK_I64)`
-/// pattern returns the lower 48 bits, which is the correct
-/// `*StringHeader` for heap strings (STRING_TAG = 0x7FFF) but is
-/// **garbage** for short-string-optimization (SSO) values
-/// (SHORT_STRING_TAG = 0x7FF9), whose lower 48 bits encode the inline
-/// length + bytes. Any consumer that dereferences the result —
-/// `js_string_concat`, `js_string_equals`, `js_string_to_lower_case`,
-/// the on-the-wire StringHeader length field, etc. — segfaults at a
-/// pseudo-random address built from the inline payload bytes.
-///
-/// Issue #214: `string[]` element loads (e.g. `JSON.parse('["hello"]')[0]`)
-/// returned SSO bits, then `arr[0] + "x"` / `arr[0] === "hello"` /
-/// `arr[0].toUpperCase()` segfaulted on the inline mask. This helper
-/// routes through `js_get_string_pointer_unified`, which materializes
-/// SSO values to a real heap StringHeader (one allocation per SSO unbox)
-/// while preserving the heap-string fast path internally.
-pub(crate) fn unbox_str_handle(blk: &mut LlBlock, boxed: &str) -> String {
-    blk.call(I64, "js_get_string_pointer_unified", &[(DOUBLE, boxed)])
-}
-
-/// Lower one of the scalar URL getters (`url.href`, `url.pathname`, …).
-/// Each runtime entry takes a raw `*mut ObjectHeader` and returns an
-/// already NaN-boxed f64 string, so the caller only has to unbox the
-/// URL handle.
-fn lower_url_string_getter(ctx: &mut FnCtx<'_>, url: &Expr, runtime_fn: &str) -> Result<String> {
-    let v = lower_expr(ctx, url)?;
-    let obj_ptr = unbox_to_i64(ctx.block(), &v);
-    Ok(ctx.block().call(DOUBLE, runtime_fn, &[(I64, &obj_ptr)]))
-}
-
-/// Lower an object literal `{ k1: v1, k2: v2, … }`.
-///
-/// Pattern:
-/// ```llvm
-/// %obj = call i64 @js_object_alloc(i32 0, i32 N)   ; class_id=0, field_count=N
-/// ; for each (key, value):
-/// %k_box = load double, ptr @.str.K.handle           ; interned key
-/// %k_bits = bitcast double %k_box to i64
-/// %k_handle = and i64 %k_bits, 281474976710655        ; POINTER_MASK_I64
-/// %v = <lower value expression>                       ; double
-/// call void @js_object_set_field_by_name(i64 %obj, i64 %k_handle, double %v)
-/// %boxed = call double @js_nanbox_pointer(i64 %obj)
-/// ```
-///
-/// Field names are interned via the StringPool, so the same key across
-/// multiple object literals shares one global string allocation.
-/// `class_id=0` is the anonymous-object class. The runtime allocates at
-/// least 8 inline field slots regardless of `field_count` to prevent
-/// buffer overflow on later set_field calls
-/// (see `crates/perry-runtime/src/object.rs:500`).
-fn lower_object_literal(ctx: &mut FnCtx<'_>, props: &[(String, Expr)]) -> Result<String> {
-    let field_count = props.len() as u32;
-    let zero_str = "0".to_string();
-    let n_str = field_count.to_string();
-
-    // Fast path: no closure-with-`this` props. Use the shape-cache allocator
-    // and write fields by INDEX — this skips the per-field linear key-search
-    // done by `js_object_set_field_by_name`. Cuts ~10ns per field on the hot
-    // path (and saves the keys_array realloc when `getDetailedIdType`-style
-    // returns are evaluated 10k×/round). Closure-with-`this` props still
-    // need the by-name path because `this_patches` populates them post-build
-    // via `js_closure_set_capture_f64`, which assumes the key is already in
-    // keys_array — fine here since the shape allocator pre-populates it.
-    let any_method_closure = props.iter().any(|(_, v)| {
-        matches!(
-            v,
-            Expr::Closure {
-                captures_this: true,
-                ..
-            }
-        )
-    });
-
-    if !any_method_closure && field_count > 0 {
-        // Build packed keys "k1\0k2\0…" interned in the StringPool (shared
-        // across all literals with the same key set + order).
-        let mut packed_keys = String::new();
-        for (k, _) in props {
-            packed_keys.push_str(k);
-            packed_keys.push('\0');
-        }
-        let keys_idx = ctx.strings.intern(&packed_keys);
-        let keys_entry = ctx.strings.entry(keys_idx);
-        let keys_global = format!("@{}", keys_entry.bytes_global);
-        let keys_len_str = keys_entry.byte_len.to_string();
-
-        // Stable shape_id derived from the packed-keys bytes. SHAPE_INLINE_CACHE
-        // is a 256-slot direct-mapped array; collisions fall through to the
-        // overflow HashMap, so any deterministic non-zero u32 works. FNV-1a
-        // is fast, dependency-free, and well-distributed across small inputs.
-        let mut shape_id: u32 = 0x811c9dc5;
-        for b in packed_keys.as_bytes() {
-            shape_id ^= *b as u32;
-            shape_id = shape_id.wrapping_mul(0x01000193);
-        }
-        if shape_id == 0 {
-            // shape_cache_get treats shape_id == 0 as "empty slot"; bump to 1.
-            shape_id = 1;
-        }
-        let shape_id_str = shape_id.to_string();
-
-        let obj_handle = ctx.block().call(
-            I64,
-            "js_object_alloc_with_shape",
-            &[
-                (I32, &shape_id_str),
-                (I32, &n_str),
-                (PTR, &keys_global),
-                (I32, &keys_len_str),
-            ],
-        );
-
-        for (i, (_, value_expr)) in props.iter().enumerate() {
-            let v = lower_expr(ctx, value_expr)?;
-            let idx_str = i.to_string();
-            // Issue #448: the runtime `js_object_set_field` takes its
-            // value as `JSValue` (`#[repr(transparent)] u64`), which the
-            // System V / AArch64 / Win64 ABIs all pass in a *general*-
-            // purpose register. The lowered NaN-box `v` is a `double`,
-            // which the same ABIs pass in a *floating-point* register.
-            // Without the bitcast the call sent the value in xmm0 / d0
-            // while Rust read garbage from rdx / x2, so generator iter
-            // objects (`{next, return, throw}` literals built via the
-            // shape-cache fast path) read back closure-typed fields as
-            // `0` — and the resulting `__iter.next()` dispatch never
-            // returned a real iter-result, so `for…of` over a class
-            // implementing `*[Symbol.iterator]()` hung forever
-            // allocating empty results.
-            let blk = ctx.block();
-            let v_bits = blk.bitcast_double_to_i64(&v);
-            blk.call_void(
-                "js_object_set_field",
-                &[(I64, &obj_handle), (I32, &idx_str), (I64, &v_bits)],
-            );
-        }
-        return Ok(nanbox_pointer_inline(ctx.block(), &obj_handle));
-    }
-
-    let obj_handle = ctx
-        .block()
-        .call(I64, "js_object_alloc", &[(I32, &zero_str), (I32, &n_str)]);
-
-    // Track `(closure_value_double, reserved_this_slot_idx)` for each
-    // method closure that needs `this` patched after the object is
-    // fully built. Enables `calc.add(n) { this.value = ... }`.
-    let mut this_patches: Vec<(String, u32)> = Vec::new();
-
-    for (key, value_expr) in props {
-        let key_idx = ctx.strings.intern(key);
-        let key_handle_global = format!("@{}", ctx.strings.entry(key_idx).handle_global);
-
-        if let Expr::Closure {
-            params: cparams,
-            body: cbody,
-            captures: ccaps,
-            captures_this: true,
-            ..
-        } = value_expr
-        {
-            let auto_caps = compute_auto_captures(ctx, cparams, cbody, ccaps);
-            let this_idx = auto_caps.len() as u32;
-
-            let v = lower_expr(ctx, value_expr)?;
-            this_patches.push((v.clone(), this_idx));
-
-            let blk = ctx.block();
-            let key_box = blk.load(DOUBLE, &key_handle_global);
-            let key_bits = blk.bitcast_double_to_i64(&key_box);
-            let key_raw = blk.and(I64, &key_bits, POINTER_MASK_I64);
-            blk.call_void(
-                "js_object_set_field_by_name",
-                &[(I64, &obj_handle), (I64, &key_raw), (DOUBLE, &v)],
-            );
-            continue;
-        }
-
-        let v = lower_expr(ctx, value_expr)?;
-        let blk = ctx.block();
-        let key_box = blk.load(DOUBLE, &key_handle_global);
-        let key_bits = blk.bitcast_double_to_i64(&key_box);
-        let key_raw = blk.and(I64, &key_bits, POINTER_MASK_I64);
-        blk.call_void(
-            "js_object_set_field_by_name",
-            &[(I64, &obj_handle), (I64, &key_raw), (DOUBLE, &v)],
-        );
-    }
-
-    // Patch each method closure's reserved `this` slot with the object
-    // pointer (NaN-boxed). Done AFTER all fields are set so every
-    // method sees the fully-initialized object.
-    if !this_patches.is_empty() {
-        let blk = ctx.block();
-        let obj_tagged = {
-            let tagged = blk.or(I64, &obj_handle, crate::nanbox::POINTER_TAG_I64);
-            blk.bitcast_i64_to_double(&tagged)
-        };
-        for (closure_val, this_idx) in &this_patches {
-            let bits = blk.bitcast_double_to_i64(closure_val);
-            let closure_handle = blk.and(I64, &bits, POINTER_MASK_I64);
-            let idx_str = this_idx.to_string();
-            blk.call_void(
-                "js_closure_set_capture_f64",
-                &[
-                    (I64, &closure_handle),
-                    (I32, &idx_str),
-                    (DOUBLE, &obj_tagged),
-                ],
-            );
-        }
-    }
-
-    Ok(nanbox_pointer_inline(ctx.block(), &obj_handle))
-}
-
-/// Lower an array literal `[a, b, c, …]`.
-///
-/// Fast path: element expressions are lowered first (any allocations
-/// inside elements complete before we claim the arena bump slot for the
-/// outer array), then for small literals (≤ 16 elements) we emit inline
-/// bump-allocator IR — the same pattern `new ClassName()` uses when
-/// `class_keys_globals` is populated. No extern call on the hot path:
-/// a load of the per-function arena state, a bump-pointer check, one i64
-/// store for the packed GcHeader, one i64 store for the packed ArrayHeader
-/// (length and capacity share the same 8 bytes), and N `store double, ptr`
-/// for the elements. The slow path (block overflow) calls
-/// `js_inline_arena_slow_alloc`.
-///
-/// For N > 16 we fall back to the extern `js_array_alloc_literal` — the
-/// inline path emits per-literal IR that's cheap at small N but grows with
-/// each element store, so large literals benefit more from a compact call.
-///
-/// GC safety: the array header is written after the bump commits, so any
-/// GC observing the partially-written arena block sees either a not-yet-
-/// allocated slot (offset hasn't advanced past the `fits` check) or a
-/// header with `length == capacity` and uninitialized elements. No
-/// allocator call runs between the header write and the element stores,
-/// so GC can't run in that window. Element expressions with their own
-/// allocations lower to SSA values pinned by conservative stack scanning.
-fn lower_array_literal(ctx: &mut FnCtx<'_>, elements: &[Expr]) -> Result<String> {
-    let n = elements.len();
-
-    // Empty literal: no elements to worry about, keep the simple path.
-    if n == 0 {
-        let arr = ctx.block().call(I64, "js_array_alloc", &[(I32, "0")]);
-        return Ok(nanbox_pointer_inline(ctx.block(), &arr));
-    }
-
-    // Evaluate all element expressions *before* allocating. This keeps each
-    // value in an SSA register (spilled to stack if needed; reachable by the
-    // conservative stack scanner) so nested allocations inside element
-    // expressions don't see a half-initialized outer array.
-    let mut vals = Vec::with_capacity(n);
-    for value_expr in elements {
-        vals.push(lower_expr(ctx, value_expr)?);
-    }
-
-    // Inline bump-allocator path for small literals. Size threshold matches
-    // `MAX_SCALAR_ARRAY_LEN` in collectors.rs so every candidate the escape
-    // pass rejects can still benefit from the inline alloc.
-    const INLINE_MAX_ELEMENTS: usize = 16;
-    if n <= INLINE_MAX_ELEMENTS {
-        // Layout constants — must match `ArrayHeader` in array.rs and
-        // `GcHeader` in gc.rs. Duplicated here because codegen emits raw
-        // byte offsets; the runtime declarations are authoritative.
-        const GC_HEADER_SIZE: u64 = 8;
-        const ARRAY_HEADER_SIZE: u64 = 8;
-        const ELEMENT_SIZE: u64 = 8;
-        const GC_TYPE_ARRAY: u64 = 1;
-        const GC_FLAG_ARENA: u64 = 0x02;
-
-        let total_size = GC_HEADER_SIZE + ARRAY_HEADER_SIZE + (n as u64) * ELEMENT_SIZE;
-        let total_size_str = total_size.to_string();
-
-        // Lazy per-function slot for the arena state pointer. Reused for
-        // `new ClassName()` inline allocs; first one to hit creates it.
-        let arena_state_slot = if let Some(slot) = ctx.arena_state_slot.clone() {
-            slot
-        } else {
-            let slot = ctx.func.entry_init_call_ptr("js_inline_arena_state");
-            ctx.arena_state_slot = Some(slot.clone());
-            slot
-        };
-
-        // Load state + compute bump check. `total_size` is always a
-        // multiple of 8, every prior alloc rounds offset to 8, and blocks
-        // start 8-aligned, so no align-up step is needed.
-        let blk = ctx.block();
-        let state_ptr = blk.load(PTR, &arena_state_slot);
-        let offset_field_ptr = blk.gep(I8, &state_ptr, &[(I64, "8")]);
-        let offset_val = blk.load(I64, &offset_field_ptr);
-        let aligned_off = offset_val.clone();
-        let new_offset = blk.add(I64, &aligned_off, &total_size_str);
-        let size_field_ptr = blk.gep(I8, &state_ptr, &[(I64, "16")]);
-        let size_val = blk.load(I64, &size_field_ptr);
-        let fits = blk.icmp_ule(I64, &new_offset, &size_val);
-
-        let fast_idx = ctx.new_block("arrlit.fast");
-        let slow_idx = ctx.new_block("arrlit.slow");
-        let merge_idx = ctx.new_block("arrlit.merge");
-        let fast_label = ctx.block_label(fast_idx);
-        let slow_label = ctx.block_label(slow_idx);
-        let merge_label = ctx.block_label(merge_idx);
-
-        ctx.block().cond_br(&fits, &fast_label, &slow_label);
-
-        // Fast path: commit the bump, compute `data + offset`.
-        ctx.current_block = fast_idx;
-        let blk = ctx.block();
-        blk.store(I64, &new_offset, &offset_field_ptr);
-        let data_ptr = blk.load(PTR, &state_ptr);
-        let raw_fast = blk.gep(I8, &data_ptr, &[(I64, &aligned_off)]);
-        let fast_pred_label = blk.label.clone();
-        blk.br(&merge_label);
-
-        // Slow path: call the runtime slow-alloc (same one used by the
-        // inline `new` path). Returns a fresh raw pointer (inclusive of
-        // GcHeader space).
-        ctx.current_block = slow_idx;
-        let raw_slow = ctx.block().call(
-            PTR,
-            "js_inline_arena_slow_alloc",
-            &[(PTR, &state_ptr), (I64, &total_size_str), (I64, "8")],
-        );
-        let slow_pred_label = ctx.block().label.clone();
-        ctx.block().br(&merge_label);
-
-        // Merge: phi the raw pointer and write everything.
-        ctx.current_block = merge_idx;
-        let blk = ctx.block();
-        let raw = blk.phi(
-            PTR,
-            &[(&raw_fast, &fast_pred_label), (&raw_slow, &slow_pred_label)],
-        );
-
-        // Packed GcHeader (bits 0..7 obj_type, 8..15 gc_flags, 16..31
-        // _reserved, 32..63 size).
-        let gc_packed: u64 = GC_TYPE_ARRAY | (GC_FLAG_ARENA << 8) | (total_size << 32);
-        blk.store(I64, &gc_packed.to_string(), &raw);
-
-        // Packed ArrayHeader at raw+8 (length low 32 / capacity high 32).
-        let arr_header_addr = blk.gep(I8, &raw, &[(I64, "8")]);
-        let arr_header_packed = (n as u64) | ((n as u64) << 32);
-        blk.store(I64, &arr_header_packed.to_string(), &arr_header_addr);
-
-        // Elements at raw+16 + i*8.
-        for (i, v) in vals.iter().enumerate() {
-            let offset = (16 + i * 8).to_string();
-            let elem_ptr = blk.gep_inbounds(I8, &raw, &[(I64, &offset)]);
-            blk.store(DOUBLE, v, &elem_ptr);
-        }
-
-        // User pointer = raw + GC_HEADER_SIZE. This is the same address
-        // `js_array_alloc_literal` returns and that the rest of the
-        // codegen expects (ArrayHeader at offset 0, elements at offset 8).
-        let user_ptr = blk.gep(I8, &raw, &[(I64, "8")]);
-        let user_ptr_as_i64 = blk.ptrtoint(&user_ptr, I64);
-        return Ok(nanbox_pointer_inline(ctx.block(), &user_ptr_as_i64));
-    }
-
-    // Fallback for N > INLINE_MAX_ELEMENTS: keep the extern call + N inline
-    // stores. Thin-LTO already inlines this call into user IR, so the cost
-    // is ~1 inlined arena bump plus some LLVM churn around the arg pack.
-    let cap_str = n.to_string();
-    let arr = ctx
-        .block()
-        .call(I64, "js_array_alloc_literal", &[(I32, &cap_str)]);
-
-    let arr_ptr = ctx.block().inttoptr(I64, &arr);
-    for (i, v) in vals.iter().enumerate() {
-        let offset = (8 + i * 8).to_string();
-        let elem_ptr = ctx.block().gep_inbounds(I8, &arr_ptr, &[(I64, &offset)]);
-        ctx.block().store(DOUBLE, v, &elem_ptr);
-    }
-
-    Ok(nanbox_pointer_inline(ctx.block(), &arr))
-}
-
-/// Inline fast-path lowering for `local_arr[i] = v`.
-///
-/// Compiles to:
-///
-/// ```text
-///   <current>:
-///     %arr_handle = unbox(arr_box)
-///     %length = load i32, ptr @ arr_handle+0
-///     %in_bounds = icmp ult %idx_i32, %length
-///     br i1 %in_bounds, label %fast_inbounds, label %check_capacity
-///
-///   fast_inbounds:
-///     ; element_ptr = arr_handle + 8 + idx*8
-///     store double %v, ptr %element_ptr
-///     br merge
-///
-///   check_capacity:
-///     %capacity = load i32, ptr @ arr_handle+4
-///     %within_cap = icmp ult %idx_i32, %capacity
-///     br i1 %within_cap, label %extend_inline, label %realloc
-///
-///   extend_inline:
-///     store double %v, ptr %element_ptr
-///     %new_len = add i32 %idx, 1
-///     store i32 %new_len, ptr @ arr_handle+0
-///     br merge
-///
-///   realloc:
-///     %new_handle = call i64 @js_array_set_f64_extend(...)
-///     %new_box = nanbox_pointer_inline(new_handle)
-///     store double %new_box, ptr %local_slot
-///     br merge
-///
-///   merge:
-///     <continues here>
-/// ```
-///
-/// The first two paths are pure inline IR — no function calls, no extra
-/// memory loads. The third path only fires when the array actually has
-/// to grow (~17 times for a 100K-element build with doubling growth).
-fn lower_index_set_fast(
-    ctx: &mut FnCtx<'_>,
-    arr_box: &str,
-    idx_double: &str,
-    val_double: &str,
-    local_id: u32,
-) -> Result<()> {
-    // Capture the local slot for the realloc path.
-    let slot = ctx
-        .locals
-        .get(&local_id)
-        .ok_or_else(|| anyhow!("IndexSet: local {} not in scope", local_id))?
-        .clone();
-
-    // Unbox the array pointer.
-    let blk = ctx.block();
-    let arr_bits = blk.bitcast_double_to_i64(arr_box);
-    let arr_handle = blk.and(I64, &arr_bits, POINTER_MASK_I64);
-    let idx_i32 = blk.fptosi(DOUBLE, idx_double, I32);
-
-    // Issue #233: detect FORWARDED arrays (post-grow stale pointers
-    // from async-fn parameter handoff) and route to the realloc slow
-    // path. The slow path's `js_array_set_f64_extend` →
-    // `clean_arr_ptr_mut` follows the forwarding chain and writes
-    // into the live new array. Without this guard, length+capacity
-    // read at offsets 0/4 would be the lower 32 bits of the
-    // forwarding pointer (garbage) and the inline element store at
-    // arr+8+idx*8 would corrupt unrelated memory.
-    let gc_flags_addr = blk.sub(I64, &arr_handle, "7");
-    let gc_flags_ptr = blk.inttoptr(I64, &gc_flags_addr);
-    let gc_flags = blk.load(I8, &gc_flags_ptr);
-    let fwd_bits = blk.and(I8, &gc_flags, "128"); // GC_FLAG_FORWARDED
-    let is_fwd = blk.icmp_ne(I8, &fwd_bits, "0");
-
-    let fwd_idx = ctx.new_block("idxset.fwd");
-    let nofwd_idx = ctx.new_block("idxset.nofwd");
-    let inbounds_idx = ctx.new_block("idxset.inbounds");
-    let check_cap_idx = ctx.new_block("idxset.check_cap");
-    let extend_inline_idx = ctx.new_block("idxset.extend_inline");
-    let realloc_idx = ctx.new_block("idxset.realloc");
-    let merge_idx = ctx.new_block("idxset.merge");
-
-    let fwd_label = ctx.block_label(fwd_idx);
-    let nofwd_label = ctx.block_label(nofwd_idx);
-    let inbounds_label = ctx.block_label(inbounds_idx);
-    let check_cap_label = ctx.block_label(check_cap_idx);
-    let extend_inline_label = ctx.block_label(extend_inline_idx);
-    let realloc_label = ctx.block_label(realloc_idx);
-    let merge_label = ctx.block_label(merge_idx);
-
-    ctx.block().cond_br(&is_fwd, &fwd_label, &nofwd_label);
-
-    // FORWARDED branch: same shape as the realloc slow path —
-    // js_array_set_f64_extend handles forwarding via clean_arr_ptr.
-    ctx.current_block = fwd_idx;
-    {
-        let blk = ctx.block();
-        let new_handle = blk.call(
-            I64,
-            "js_array_set_f64_extend",
-            &[(I64, &arr_handle), (I32, &idx_i32), (DOUBLE, val_double)],
-        );
-        let new_box = nanbox_pointer_inline(blk, &new_handle);
-        blk.store(DOUBLE, &new_box, &slot);
-        let val_bits = blk.bitcast_double_to_i64(val_double);
-        emit_write_barrier_slot_on_block(blk, &arr_handle, "0", &val_bits);
-        blk.br(&merge_label);
-    }
-
-    ctx.current_block = nofwd_idx;
-    // Load length from offset 0 (null-guarded).
-    let length = ctx.block().safe_load_i32_from_ptr(&arr_handle);
-    let in_bounds = ctx.block().icmp_ult(I32, &idx_i32, &length);
-    ctx.block()
-        .cond_br(&in_bounds, &inbounds_label, &check_cap_label);
-
-    // Helper: compute element_ptr = arr_ptr + 8 + idx*8 and emit a store.
-    fn store_element(
-        blk: &mut LlBlock,
-        arr_handle: &str,
-        idx_i32: &str,
-        val_double: &str,
-    ) -> String {
-        let idx_i64 = blk.zext(I32, idx_i32, I64);
-        let byte_offset = blk.shl(I64, &idx_i64, "3"); // *8
-        let with_header = blk.add(I64, &byte_offset, "8"); // +8 for header
-        let element_addr = blk.add(I64, arr_handle, &with_header);
-        let element_ptr = blk.inttoptr(I64, &element_addr);
-        blk.store(DOUBLE, val_double, &element_ptr);
-        element_addr
-    }
-
-    // FASTEST: in-bounds path. Store directly, jump to merge.
-    ctx.current_block = inbounds_idx;
-    {
-        let blk = ctx.block();
-        let element_addr = store_element(blk, &arr_handle, &idx_i32, val_double);
-        let val_bits = blk.bitcast_double_to_i64(val_double);
-        emit_write_barrier_slot_on_block(blk, &arr_handle, &element_addr, &val_bits);
-        blk.br(&merge_label);
-    }
-
-    // MEDIUM: idx >= length but < capacity. Store + bump length.
-    ctx.current_block = check_cap_idx;
-    let capacity = {
-        let blk = ctx.block();
-        // Load capacity from offset 4 — we need a typed pointer that
-        // points 4 bytes into the array header. Use inttoptr after add.
-        let cap_addr = blk.add(I64, &arr_handle, "4");
-        let cap_ptr = blk.inttoptr(I64, &cap_addr);
-        blk.load(I32, &cap_ptr)
-    };
-    let within_cap = ctx.block().icmp_ult(I32, &idx_i32, &capacity);
-    ctx.block()
-        .cond_br(&within_cap, &extend_inline_label, &realloc_label);
-
-    ctx.current_block = extend_inline_idx;
-    {
-        let blk = ctx.block();
-        let element_addr = store_element(blk, &arr_handle, &idx_i32, val_double);
-        // Bump length: store idx+1 to arr_ptr+0.
-        let new_len = blk.add(I32, &idx_i32, "1");
-        let len_ptr = blk.inttoptr(I64, &arr_handle); // length is at offset 0
-        blk.store(I32, &new_len, &len_ptr);
-        let val_bits = blk.bitcast_double_to_i64(val_double);
-        emit_write_barrier_slot_on_block(blk, &arr_handle, &element_addr, &val_bits);
-        blk.br(&merge_label);
-    }
-
-    // SLOW: realloc needed. Call the runtime, write new ptr to local.
-    ctx.current_block = realloc_idx;
-    {
-        let blk = ctx.block();
-        let new_handle = blk.call(
-            I64,
-            "js_array_set_f64_extend",
-            &[(I64, &arr_handle), (I32, &idx_i32), (DOUBLE, val_double)],
-        );
-        let new_box = nanbox_pointer_inline(blk, &new_handle);
-        blk.store(DOUBLE, &new_box, &slot);
-        let val_bits = blk.bitcast_double_to_i64(val_double);
-        emit_write_barrier_slot_on_block(blk, &arr_handle, "0", &val_bits);
-        blk.br(&merge_label);
-    }
-
-    ctx.current_block = merge_idx;
-    Ok(())
 }
